cleanup unused files and rename publish_files.py to ingest_files.py
